Begin adding commands from stock lib
The ril.h might need some work on the block of added defines. I believe
this brings the ril code "up-to-date" if you will with all the commands
seen in the decompiled code. So there should be nothing missing and
therefore things like emergency calling should work. And nothing was
changed to the core code so it should work as it did before.